+++ .gitignore
@@ -18,6 +18,7 @@
 *.ear
 *.zip
 *.tar.gz
+*.gz
 *.rar
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
