+++ build.gradle
@@ -26,6 +26,9 @@ allprojects {
     apply plugin: "io.spring.dependency-management"
     dependencyManagement {
         dependencies {
+            imports {
+                mavenBom("org.springframework.boot:spring-boot-dependencies:2.3.1.RELEASE")
+            }
             dependency("com.google.guava:guava:${Versions.guava}")
         }
     }
