+++ hw04-GC/src/main/java/ru/otus/torwel/Benchmark.java
@@ -17,14 +17,14 @@ public class Benchmark implements BenchmarkMBean {
         for (int idx = 0; idx < loopCounter; idx++) {
             int local = size;
 
-            System.out.println("Creating objects...");
+//            System.out.println("Creating objects...");
             for (int i = 0; i < local; i++) {
                 list.add(new String(new char[0]));
 //                if (i % 1000 == 0)
 //                    System.out.println("add " + i);
             }
 
-            System.out.println("Removing objects...");
+//            System.out.println("Removing objects...");
             var lstSize = list.size();
             for (int i = lstSize - 1; i >= lstSize - local/4*3; i--) {
                 list.remove(i);
@@ -32,16 +32,8 @@ public class Benchmark implements BenchmarkMBean {
 //                    System.out.println("remove " + i);
             }
 
-/*
-            Object[] array = new Object[local];
-            for (int i = 0; i < local; i++) {
-                array[i] = new String(new char[0]);
-            }
-*/
 
-            System.out.println("loop: " + idx + "\tlstSize: " + lstSize);
-            System.out.println("Maximum duration STW: " + Launch.listener.getMaxDurationSTW());
-            System.out.println("Total duration STWs: " + Launch.listener.getTotalDurationSTW());
+//            System.out.println("loop: " + idx + "\tlstSize: " + lstSize);
             Thread.sleep(10); //Label_1
         }
     }
