+++ hw04-GC/src/main/resources/log4j2.properties
@@ -1,0 +1,36 @@
+status = error
+name = PropertiesConfig
+
+#Make sure to change log file path as per your need
+property.filename = C:\\logs\\info.log
+
+filters = threshold
+
+filter.threshold.type = ThresholdFilter
+filter.threshold.level = info
+
+appenders = rolling
+
+appender.rolling.type = RollingFile
+appender.rolling.name = RollingFile
+appender.rolling.fileName = ${filename}
+appender.rolling.filePattern = debug-backup-%d{MM-dd-yy-HH-mm-ss}-%i.log.gz
+appender.rolling.layout.type = PatternLayout
+appender.rolling.layout.pattern = %d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
+appender.rolling.policies.type = Policies
+appender.rolling.policies.time.type = TimeBasedTriggeringPolicy
+appender.rolling.policies.time.interval = 1
+appender.rolling.policies.time.modulate = true
+appender.rolling.policies.size.type = SizeBasedTriggeringPolicy
+appender.rolling.policies.size.size=10MB
+appender.rolling.strategy.type = DefaultRolloverStrategy
+appender.rolling.strategy.max = 20
+
+loggers = rolling
+
+#Make sure to change the package structure as per your application
+
+logger.rolling.name = ru.otus.torwel
+logger.rolling.level = info
+logger.rolling.additivity = true
+logger.rolling.appenderRef.rolling.ref = RollingFile
